pyproject.toml
[tool.black]
line-length = 79

[project]
name = "varwg"
version = "1.4.0"
description = "A Vector Autoregressive Weather Generator"
readme = "README.md"
authors = [
  { name = "iskur", email = "32363199+iskur@users.noreply.github.com" }
]
license = "BSD-3-Clause"
license-files = ["LICENSE.txt"]
requires-python = ">=3.13"
dependencies = [
    "bottleneck>=1.5.0",
    "cartopy>=0.24.1",
    "dill>=0.4.0",
    "matplotlib>=3.10.3",
    "numexpr>=2.11.0",
    "pandas>=2.3.0",
    "scipy>=1.15.3",
    "timezonefinder>=6.5.9",
    "tqdm>=4.67.1",
    "xarray>=2025.4.0",
]

[build-system]
requires = ["setuptools>=61", "cython", "numpy"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
varwg = ["*.met", "test_data/*.nc"]

[dependency-groups]
dev = [
  "cython>=3.1.1",
  "docutils>=0.21.2",
  "ipython>=9.3.0",
  "numpydoc>=1.8.0",
  "pooch>=1.8.2",
  "restructuredtext-lint>=1.4.0",
  "tox>=4.27.0",
  "twine>=6.2.0",
]
test = [
  "pytest>=8.4.0",
  "cython>=3.1.1",
  "setuptools>=61",
  "pooch>=1.8.2",
  "dwd_opendata @ git+https://github.com/iskur/dwd_opendata.git",
]

[tool.pytest.ini_options]
addopts = "--pdb"
markers = [
    "network: tests that require network access (deselect with '-m \"not network\"')",
]

[tool.cibuildwheel]
# Build for common platforms
build = "cp313-*"
skip = "pp* *-musllinux_*"

# Skip testing during build to avoid installing heavy dependencies (cartopy->pyproj->proj)
# Extensions are validated by successful compilation; smoke tests run in CI workflow
test-skip = "*"

# Platform-specific settings
[tool.cibuildwheel.linux]
archs = ["x86_64"]
before-build = "pip install cython numpy setuptools"
# Note: aarch64 requires QEMU emulation and significantly increases build time
# Can be added later if ARM Linux support is needed

[tool.cibuildwheel.windows]
archs = ["AMD64"]
before-build = "pip install cython numpy setuptools"

[tool.cibuildwheel.macos]
archs = ["x86_64", "arm64"]
before-build = "pip install cython numpy setuptools"
